rename: 프롬프트 조회 API 요청 함수 파일명 변경
*getPromptList → prompts
